requirements.txt
python-dotenv
streamlit
feedparser
arxiv
pymupdf

# Vector RAG stack
langchain
langchain-community
langchain-huggingface
sentence-transformers
langchain-chroma
chromadb

# Graph RAG stack
neo4j

